Add support for Rails 6.1.
I've removed the coercing of the default value for an enum column to a
symbol in the column definition. This causes an explosion on Rails 6.1.
It is not actually needed, since the value will be typecast before
assigning and a new record has symbol values for any enum columns with
defaults defined.